requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
et-xmlfile==1.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.27.1
six==1.16.0
urllib3==1.26.14
yarl==1.8.2
